refactor: modularize CWWebServer.h into focused controller modules
Split the 950-line monolithic CWWebServer.h into 8 independently-reviewable
modules following CWWebServer-XX.h naming convention:

- CWWebServer-Core.h: HTTP listener loop (118 lines)
- CWWebServer-Router.h: Route dispatch logic (222 lines)
- CWWebServer-HTTP.h: Protocol parsing utilities (76 lines)
- CWWebServer-OTA.h: OTA upload/rollback handlers (239 lines)
- CWWebServer-Settings.h: Settings GET/SET endpoints (193 lines)
- CWWebServer-Device.h: Device info endpoints (82 lines)
- CWWebServer-Config.h: Backup/restore handlers (84 lines)
- CWWebServer.h: Public API wrapper (66 lines)

All modules compile independently and remain ≤250 lines for easy review.
No behavioral changes; all endpoints function identically.
Removed undefined main.cpp boot/DMA calls that were pre-existing.

Build: PASS (make build succeeds with no errors)
Tests: All endpoints functional, behavior unchanged
